+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 3.1.0a2 (unreleased)
 --------------------
 
+- Transforms the display of project themes to links
+  [vpiret]
+
 - Add StateProgress view
   [vpiret]